Add testnet deployments

+++ contracts/GenericERC20.sol
@@ -1,42 +1,50 @@
 pragma solidity ^0.5.16;
 
 contract GenericERC20 {
-
     string public name = "ERC20Basic";
     string public symbol = "USDC";
-    uint8 public decimals = 6;  
+    uint8 public decimals = 6;
 
-
-    event Approval(address indexed tokenOwner, address indexed spender, uint tokens);
-    event Transfer(address indexed from, address indexed to, uint tokens);
-
+    event Approval(
+        address indexed tokenOwner,
+        address indexed spender,
+        uint256 tokens
+    );
+    event Transfer(address indexed from, address indexed to, uint256 tokens);
 
     mapping(address => uint256) balances;
 
-    mapping(address => mapping (address => uint256)) allowed;
-    
+    mapping(address => mapping(address => uint256)) allowed;
+
     uint256 totalSupply_;
 
     using SafeMath for uint256;
 
-
-   constructor(uint256 total, string memory _name, string memory _symbol, uint8 _decimals) public {  
-	totalSupply_ = total;
-	balances[msg.sender] = totalSupply_;
-    name = _name;
-    symbol = _symbol;
-    decimals = _decimals;
-    }  
+    constructor(
+        uint256 total,
+        string memory _name,
+        string memory _symbol,
+        uint8 _decimals
+    ) public {
+        totalSupply_ = total;
+        balances[msg.sender] = totalSupply_;
+        name = _name;
+        symbol = _symbol;
+        decimals = _decimals;
+    }
 
     function totalSupply() public view returns (uint256) {
-	return totalSupply_;
+        return totalSupply_;
     }
-    
-    function balanceOf(address tokenOwner) public view returns (uint) {
+
+    function balanceOf(address tokenOwner) public view returns (uint256) {
         return balances[tokenOwner];
     }
 
-    function transfer(address receiver, uint numTokens) public returns (bool) {
+    function transfer(address receiver, uint256 numTokens)
+        public
+        returns (bool)
+    {
         require(numTokens <= balances[msg.sender]);
         balances[msg.sender] = balances[msg.sender].sub(numTokens);
         balances[receiver] = balances[receiver].add(numTokens);
@@ -44,42 +52,53 @@ contract GenericERC20 {
         return true;
     }
 
-    function approve(address delegate, uint numTokens) public returns (bool) {
+    function approve(address delegate, uint256 numTokens)
+        public
+        returns (bool)
+    {
         allowed[msg.sender][delegate] = numTokens;
         emit Approval(msg.sender, delegate, numTokens);
         return true;
     }
 
-    function allowance(address owner, address delegate) public view returns (uint) {
+    function allowance(address owner, address delegate)
+        public
+        view
+        returns (uint256)
+    {
         return allowed[owner][delegate];
     }
 
-    function transferFrom(address owner, address buyer, uint numTokens) public returns (bool) {
-        require(numTokens <= balances[owner]);    
+    function transferFrom(
+        address owner,
+        address buyer,
+        uint256 numTokens
+    ) public returns (bool) {
+        require(numTokens <= balances[owner]);
         require(numTokens <= allowed[owner][msg.sender]);
-    
+
         balances[owner] = balances[owner].sub(numTokens);
         allowed[owner][msg.sender] = allowed[owner][msg.sender].sub(numTokens);
         balances[buyer] = balances[buyer].add(numTokens);
         emit Transfer(owner, buyer, numTokens);
         return true;
     }
-    
-    function mintTokens(address to, uint numTokens) public {
+
+    function mintTokens(address to, uint256 numTokens) public {
         balances[to] += numTokens;
         totalSupply_ += numTokens;
     }
 }
 
-library SafeMath { 
+library SafeMath {
     function sub(uint256 a, uint256 b) internal pure returns (uint256) {
-      assert(b <= a);
-      return a - b;
+        assert(b <= a);
+        return a - b;
     }
-    
+
     function add(uint256 a, uint256 b) internal pure returns (uint256) {
-      uint256 c = a + b;
-      assert(c >= a);
-      return c;
+        uint256 c = a + b;
+        assert(c >= a);
+        return c;
     }
 }
